+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- Set argument parser program name to script filename without extension for better help display
+- Support for `# Description: ...` annotation to add script description to help output
 
 ### Changed